requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
numpy>=1.24.0